kernel: move AppId to process.rs
